requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.0.9
Django==3.2.7
django-session-timeout==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.1.0
waitress==2.0.0
wcwidth==0.2.5
whitenoise==5.3.0
